pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tugastugas"
version = "0.0.1"
authors = [{name = "Vee Satayamas", email = "vsatayamas@gmail.com"}]
description = "A showcase for GraphQL, Python, and PostgreSQL"
readme = "README.md"
keywords = ["graphql"]
license = {file = "LICENSE"}
classifiers = [
	    "Programming Language :: Python :: 3",
	    ]
dependencies = [
	     "psycopg>=3.1,<3.2",
	     "alembic>=1.13",
	     "pytest-alembic>=0.11,<0.12",
	     "pytest-mock-resources[docker]",
	     "graphene-sqlalchemy==v3.0.0rc1",
	     "sqlalchemy>=2"
]

[project.urls]
homepage = "https://git.sr.ht/~veer66/tugastugas"

[tool.pylsp-mypy]
enabled = true
live_mode = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
